% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dime_data}
\alias{dime_data}
\title{Dime Dataset}
\description{
A set of donor names from the Database on Ideology, Money in Politics, and
Elections (DIME).  This dataset was used as a benchmark in the 2021 APSR
paper Adaptive Fuzzy String Matching: How to Merge Datasets with Only One
(Messy) Identifying Field by Aaron R. Kaufman and Aja Klevs, the dataset in
this package is a subset of the data from the replication archive of that
paper. The full dataset can be found in the paper's replication materials
here: \doi{10.7910/DVN/4031UL}.
}
\references{
\doi{10.7910/DVN/4031UL}
}
\author{
Adam Bonica
}
