% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_link.R
\name{em_link}
\alias{em_link}
\title{Fit a Probabilistic Matching Model using Naive Bayes + E.M.}
\usage{
em_link(X, g, tol = 10^-6, max_iter = 10^3)
}
\arguments{
\item{X}{a matrix of similarities}

\item{g}{a vector of initial guesses}

\item{tol}{tolerance in the sense of the infinity norm. i.e.}

\item{max_iter}{iterations after which the algorithm will error out if it has not converged}
}
\value{
a vector of probabilities representing the posterior probability
each record pair is a match.
}
\description{
Fit a Probabilistic Matching Model using Naive Bayes + E.M.
}
