% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsh_properties.R
\name{euclidean_curve}
\alias{euclidean_curve}
\title{Plot S-Curve for a LSH with given hyperparameters}
\usage{
euclidean_curve(n_bands, band_width, r, up_to = 100)
}
\arguments{
\item{n_bands}{The number of LSH bands calculated}

\item{band_width}{The number of hashes in each band}

\item{r}{the "r" hyperparameter used to govern the sensitivity of the hash.}

\item{up_to}{the right extent of the x axis.}
}
\value{
A plot showing the probability a pair is proposed as a match, given
the Jaccard similarity of the two items.
}
\description{
Plot S-Curve for a LSH with given hyperparameters
}
