% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsh_properties.R
\name{euclidean_probability}
\alias{euclidean_probability}
\title{Find Probability of Match Based on Similarity}
\usage{
euclidean_probability(distance, n_bands, band_width, r)
}
\arguments{
\item{distance}{the euclidian distance between the two vectors you want to
compare.}

\item{n_bands}{The number of LSH bands used in hashing.}

\item{band_width}{The number of hashes in each band.}

\item{r}{the "r" hyperparameter used to govern the sensitivity of the hash.}
}
\value{
a decimal number giving the proability that the two items will be
returned as a candidate pair from the minihash algorithm.
}
\description{
Find Probability of Match Based on Similarity
}
