% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_core.R
\name{fuzzy_join_core}
\alias{fuzzy_join_core}
\title{Perform a Fuzzy-Join With an Arbitrary Distance Metric}
\usage{
fuzzy_join_core(
  a,
  b,
  by,
  join_func,
  mode,
  block_by = NULL,
  similarity_column = NULL,
  ...
)
}
\arguments{
\item{a}{the first dataframe to join}

\item{b}{the second dataframe to join}

\item{by}{columns on which to join}

\item{join_func}{the joining function responsible for performing the join}

\item{mode}{the dplyr-style type of join you want to perform}

\item{block_by}{any columns to block (perform exact matching on) (where supported)}

\item{similarity_column}{name of the column with the distance metric in the
resulting dataframe}

\item{...}{Other parameters to be passed to the joining function}
}
\description{
Perform a Fuzzy-Join With an Arbitrary Distance Metric
}
