% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard_similarity.R
\name{hamming_distance}
\alias{hamming_distance}
\title{Calculate Hamming distance of two character vectors}
\usage{
hamming_distance(a, b, ngram_width = 2)
}
\arguments{
\item{a}{the first character vector}

\item{b}{the first character vector}
}
\value{
a vector of hamming similarities of the strings
}
\description{
Calculate Hamming distance of two character vectors
}
\examples{
hamming_distance(c("the quick brown fox","jumped over the lazy dog"),
    c("the quck bron fx","jumped over hte lazy dog"))

}
