% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamming_logical_joins.R
\name{hamming_inner_join}
\alias{hamming_inner_join}
\title{Fuzzy inner-join using minihashing}
\usage{
hamming_inner_join(
  a,
  b,
  by = NULL,
  n_bands = 100,
  band_width = 100,
  threshold = 2,
  progress = FALSE,
  clean = FALSE,
  similarity_column = NULL
)
}
\arguments{
\item{a}{the first dataframe you wish to join.}

\item{b}{the second dataframe you wish to join.}

\item{by}{a named vector indicating which columns to join on. Format should
be the same as dplyr: \code{by = c("column_name_in_df_a" = "column_name_in_df_b")}, but
two columns must be specified in each dataset (x column and y column). Specification
made with \code{dplyr::join_by()} are also accepted.}

\item{n_bands}{the number of bands used in the locality sensitive hashing
algorithm (default is 100). Use this in conjunction with the
\code{band_width} to determine the performance of the hashing.}

\item{band_width}{the length of each band used in the minihashing algorithm
(default is 8) Use this in conjunction with the \code{n_bands} to determine
the performance of the hashing.}

\item{progress}{set to \code{TRUE} to print progress}

\item{clean}{should the strings that you fuzzy join on be cleaned (coerced
to lower-case, stripped of punctuation and spaces)? Default is FALSE}

\item{similarity_column}{an optional character vector. If provided, the data
frame will contain a column with this name giving the jaccard similarity
between the two fields. Extra column will not be present if anti-joining.}
}
\value{
a tibble fuzzily-joined on the basis of the variables in \code{by.} Tries
to adhere to the same standards as the dplyr-joins, and uses the same
logical joining patterns (i.e. inner-join joins and keeps only observations in both datasets).
}
\description{
Find similar rows between two tables using the hamming distance. The hamming
distance is equal to the number characters two strings differ by, or is
equal to infinity if two strings are of different lengths
}
