% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsh_properties.R
\name{hamming_probability}
\alias{hamming_probability}
\title{Find Probability of Match Based on Similarity}
\usage{
hamming_probability(distance, input_length, n_bands, band_width)
}
\arguments{
\item{distance}{The hamming distance of the two strings you want to compare}

\item{input_length}{the length (number of characters) of the input strings
you want to calculate.}

\item{n_bands}{The number of LSH bands used in hashing.}

\item{band_width}{The number of hashes in each band.}
}
\value{
A decimal number giving the probability that the two items will be
returned as a candidate pair from the lsh algotithm.
}
\description{
Find Probability of Match Based on Similarity
}
