% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kd_logical_joins.R
\name{kd_inner_join}
\alias{kd_inner_join}
\title{Spatial Inner Join Using KD-Trees}
\usage{
kd_inner_join(a, b, by = NULL, threshold = 1)
}
\arguments{
\item{a}{the first dataframe you wish to join.}

\item{b}{the second dataframe
you wish to join.}

\item{by}{a named vector indicating which columns to join on. Format should
be the same as dplyr: \code{by = c("column_name_in_df_a" = "column_name_in_df_b")}, but
two columns must be specified in each dataset (x column and y column).}

\item{threshold}{the distance threshold below which units should be considered a match}
}
\description{
Spatial Inner Join Using KD-Trees
}
\examples{
n <- 10

X_1 <- matrix(c(seq(0,1,1/(n-1)), seq(0,1,1/(n-1))), nrow=n)
X_2 <- X_1 + .0000001

X_1 <- as.data.frame(X_1)
X_2 <- as.data.frame(X_2)

X_1$id_1 <- 1:n
X_2$id_2 <- 1:n

kd_inner_join(X_1, X_2, threshold =.00005)


}
