% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical_joins.R
\name{lsh_full_join}
\alias{lsh_full_join}
\title{Fuzzy full-join using MiniHashing}
\usage{
lsh_full_join(
  a,
  b,
  match_col = NULL,
  n_gram_width = 2,
  n_bands = 40,
  band_width = 5,
  threshold = 0.95
)
}
\arguments{
\item{a}{the first dataframe you wish to join}

\item{b}{the second dataframe you wish to join}

\item{match_col}{a named vector indicating which columns to join on}

\item{n_gram_width}{the length of the n_grams used in calculating the jaccard similarity}

\item{n_bands}{a named vector indicating which columns to join on}

\item{band_width}{the length of each band used in the minihashing algorithm (default is 5)}
}
\description{
Fuzzy full-join using MiniHashing
}
