% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsh_properties.R
\name{lsh_hyperparameters}
\alias{lsh_hyperparameters}
\title{Choose the Appropriate LSH hyperparamaters}
\usage{
lsh_hyperparameters(s1 = 0.1, s2 = 0.7, p1 = 0.001, p2 = 0.999)
}
\description{
Choose the Appropriate LSH hyperparamaters
}
