% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_lsh_join}
\alias{rust_lsh_join}
\title{Return string \code{"Hello world!"} to R.}
\usage{
rust_lsh_join(
  left_string_r,
  right_string_r,
  ngram_width,
  n_bands,
  band_size,
  threshold
)
}
\description{
Return string \code{"Hello world!"} to R.
}
