% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{package}
\name{zoomerjoin-package}
\alias{zoomerjoin}
\alias{zoomerjoin-package}
\title{zoomerjoin: Insanely fast fuzzy joins}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Zoomerjoin empowers users to fuzzily-merge dataframes with millions or tens of millions of rows in minuites with low memory usage. The package uses two locality sensitive hashing algorithms to avoid having to compare every pair of records in each dataset, resulting in fuzzy-merges that finish in linear time.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://beniaminogreen.github.io/zoomerjoin/}
  \item Report bugs at \url{https://github.com/beniaminogreen/zoomerjoin/issues/}
}

}
\author{
\strong{Maintainer}: Beniamino Green \email{beniamino.green@yale.edu}

Other contributors:
\itemize{
  \item Etienne Bacher \email{etienne.bacher@protonmail.com} [contributor]
}

}
