% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_core.R
\name{fuzzy_join_core}
\alias{fuzzy_join_core}
\title{Core Fuzzy Joining Logic}
\usage{
fuzzy_join_core(
  a,
  b,
  by,
  join_func,
  mode,
  block_by = NULL,
  similarity_column = NULL,
  ...
)
}
\arguments{
\item{a}{the first dataframe to join}

\item{b}{the second dataframe to join}

\item{by}{columns on which to join}

\item{join_func}{the joining function responsible for performing the join}

\item{mode}{the dplyr-style type of join you want to perform}

\item{...}{Other parameters to be passed to the joining function}

\item{block_on}{any columns to block (perform exact matching on) (where supported)}
}
\description{
Core Fuzzy Joining Logic
}
