% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_similarities.R
\name{hamming_distance}
\alias{hamming_distance}
\title{Calculate Hamming distance of two character vectors}
\usage{
hamming_distance(a, b)
}
\arguments{
\item{a}{the first character vector}

\item{b}{the first character vector}
}
\value{
a vector of hamming similarities of the strings
}
\description{
Calculate Hamming distance of two character vectors
}
\examples{
hamming_distance(
  c("ACGTCGATGACGTGATGCGTAGCGTA", "ACGTCGATGTGCTCTCGTCGATCTAC"),
  c("ACGTCGACGACGTGATGCGCAGCGTA", "ACGTCGATGGGGTCTCGTCGATCTAC")
)

}
