% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard_similarity.R
\name{jaccard_similarity}
\alias{jaccard_similarity}
\title{Calculate jaccard_similarity of two character vectors}
\usage{
jaccard_similarity(a, b, ngram_width = 2)
}
\description{
Calculate jaccard_similarity of two character vectors
}
