% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard_similarity.R
\name{jaccard_similarity}
\alias{jaccard_similarity}
\title{Calculate jaccard_similarity of two character vectors}
\usage{
jaccard_similarity(a, b, ngram_width = 2)
}
\arguments{
\item{a}{the first character vector}

\item{b}{the first character vector}

\item{ngram_width}{the length of the shingles / ngrams used in the
similarity calculation}
}
\value{
a vector of jaccard similarities of the strings
}
\description{
Calculate jaccard_similarity of two character vectors
}
\examples{
jaccard_similarity(c("the quick brown fox","jumped over the lazy dog"),
    c("the quck bron fx","jumped over hte lazy dog"))

}
