% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical_joins.R
\name{lsh_full_join}
\alias{lsh_full_join}
\title{Fuzzy full-join using MiniHashing}
\usage{
lsh_full_join(
  a,
  b,
  match_col = NULL,
  n_gram_width = 2,
  n_bands = 45,
  band_width = 8,
  threshold = 0.95
)
}
\arguments{
\item{a}{the first dataframe you wish to join.}

\item{b}{the second dataframe you wish to join.}

\item{match_col}{a named vector indicating which columns to join on. Format
should be the same as dplyr: \code{match_col = c("column_name_in_df_a" =
"column_name_in_df_b")}}

\item{n_gram_width}{the length of the n_grams used in calculating the
jaccard similarity. For best performance, I set this large enough that the
chance any string has a specific n_gram is low (i.e. \code{n_gram_width} = 2
or 3 when matching on first names, 5 or 6 when matching on entire
sentences).}

\item{n_bands}{the number of bands used in the minihash algorithm (default
is 40). Use this in conjunction with the \code{band_width} to determine the
performance of the hashing. The default settings are for a
(.2,.8,.001,.999)-sensitive hash i.e. that pairs with a similarity of less
than .2 have a >.1\% chance of being compared, while pairs with a similarity
of greater than .8 have a >99.9\% chance of being compared.}

\item{band_width}{the length of each band used in the minihashing algorithm
(default is 8) Use this in conjunction with the \code{n_bands} to determine
the performance of the hashing. The default settings are for a
(.2,.8,.001,.999)-sensitive hash i.e. that pairs with a similarity of less
than .2 have a >.1\% chance of being compared, while pairs with a similarity
of greater than .8 have a >99.9\% chance of being compared.}
}
\description{
Fuzzy full-join using MiniHashing
}
