% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsh_properties.R
\name{lsh_hyperparameters}
\alias{lsh_hyperparameters}
\title{Help Choose the Appropriate LSH Hyperparamaters}
\usage{
lsh_hyperparameters(s1 = 0.1, s2 = 0.7, p1 = 0.001, p2 = 0.999)
}
\arguments{
\item{s1:}{the s1 paramater (the first similaity).}

\item{s2:}{the s2 parameter (the second similarity, must be greater than s1).}

\item{p1:}{the p1 paramater (the first probability).}

\item{p2:}{the p2 parameter (the second probability, must be greater than p1).}
}
\description{
Runs a grid search to find the hyperparameters that will achieve an
(s1,s2,p1,p2)-sensitive locality sensitive hash. A locality sensitive hash
can be called (s1,s2,p1,p2)-sensitive if to strings with a similarity less
than s1 have a less than p1 chance of being compared, while two strings with
similarity s2 have a greater than p2 chance of being compared. As an
example, a (.1,.7,.001,.999)-sensitive LSH means that strings with
similarity less than .1 will have a .1\% chance of being compared, while
strings with .7 similarity have a 99.9\% chance of being compared.
}
