% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsh_properties.R
\name{lsh_probability}
\alias{lsh_probability}
\title{Find Probability of Match Based on Similarity}
\usage{
lsh_probability(similarity, n_bands, band_width)
}
\description{
Find Probability of Match Based on Similarity
}
